Fitness Center planning
A fitness center needs a Java application to manage its exercise equipment.Each piece of equipment is characterized by a UNIQUE serial number(id) and a Boolean value to check its state (if it needs regular maintenance of not).
The equipment includes both Cardio Machines and Strength machines:
-Cardio Machine : characterized by resistance level. 
-Strength training machine :characterized by a motion level and a muscle group (legs,arms,core)

To make it more generic I took the following adjustments:
The price for the cardio machine is 500 if ressistance is <10. Otherwise the price would be 200.
The price for strength training of arms or legs is 600.
If the machine works the core then it will be 800.
Also if it needs maintanance , the machine would have the cost doubled.

The application has the following functionalities:
Add a new training machine,checking its type in advance.
Showing all the available machines
Showing all machines cheaper than a giben balue using JAVA 8 STREAMS

I used Layered Architecture .
Used the following UML diagram for guidance.